% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11benchmark-package.R
\docType{package}
\name{cpp11armadillobenchmark-package}
\alias{cpp11armadillobenchmark}
\alias{cpp11armadillobenchmark-package}
\title{cpp11armadillobenchmark: Checking speed improvement/worsening with different C++ standards for cpp11}
\description{
Benchmark adapted from the official Armadillo documentation to test the speed of different C++ standards when using the armadillo4r package. The idea is to check how compiling with GCC or Clang using C++11, C++14, C++17, C++20 or C++23 affects the speed of various linear algebra operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

}
\keyword{internal}
